Release with computation of GHA, SHA und DECLINATION
This release offer three modes to calculate a position from as less complex as possible to as complex as possible. It can be used by using the Nautical Almanac or the Long Term Almanac but it can also compute the figures taken from the NA by itself. I compared different calculation and in the three cases I tested the difference was below 3 nautical miles. In my opinion, at open sea, this do not really matter. To get better results it makes sense to recalculate all of the correction values for the navigational stars. If you want you can do this by yourself.