Major Change
Added a dialog which allow the user to modify GHA, GHA correction, DECLINATION and DECLINATION correction by himself. The value are stored in Shared Preferences. That's why they are more or less persistent. There is an RESET button to get the Defaults back. Now a user can set the Values from a specific date of the Nautical Almanac and compute corrections to have as best as possible date for one year. In my opinion these figure can be good fpr 50 years as used in the Long Term Almanac.