Good Release
Nearby all of the minor bugs fixed. User interface is also intuitive. Unnecesary steps removed. To describe the way the user interface works is quite complex. Especially in the Nautical Almanac dialog because in some cases the figures for GHA SHA and DECLINATION are computed (Push PREV and NEXT or changes to Time and Date of the observed CB) but it's allowed to use figures from an Almanac and type them into the GHAAries, GHAAries+1, SHA and Decl. fields. I try to give a hint by setting GHA Aries and Aries+1 to "Green" background. Color blind can probably not use this feature. If "Hc-omput" in"Nautical Almana" dialog is not same as in the HO2HC dialog "Sextant Corrections" than the content of "Sextant Correction" is reseted and Ho become there Hc of the "Nautical Almanac" dialog. Check it by yourself.