New Release with SUN GHA and Declination
Added a dialog to offer calculation of GHA and Declination. This dialog is "fancy" but "faulty". I have to add more math to this dialog because planet earth is a sphere and the Moon is moving our planet a little bit and the planet has an elliptical orbit around the sun. The impact is a huge fault for Astro Navigators. I'm just working on "Jean Meeus Astronomical Algorithms" to make this work. Seems that there is no simple solution to have this the way the Nautical Almanace offers it. However it will take some time but I will find a proper solution.