Added standard behaviour to SUN dialog
The dialog of "Get declination of SUN" is very special. I added green and red background to calculated field. For the green ones I'm sure that the result is good (not excellent) but the red one is just in doubt. Because the calculation in this dialog depend on the Latitude and Longitude calculated at the "Simple Navigation (SUN)" dialog. If this data is false then the elevation is false and since the Laititude is calculated by the elevation than the figure is probably also false. It should be the Latitide at the given time where the SUN is at highest point. Well...I have to check this.  See source. There is a ToDo: comment.