Added Issue with TimeZone
I use some code from  John Clark Craig (SUN POSITION). It seems that in the USA the minus for longitudes is used different. That's why some of the calculations went totally wrong since I added TZ to the "Get declination of SUN" dialog.  A added some colors to this dialog. The green figures are good.  You can compare them for example at https://www.suncalc.org/#/54.2996,9.6654,12/2023.03.10/14:00/1/3 . Bearing and Elevation fit. Declination is now nearby and good in summer. The GHA is up to two degrees false. I change the code to be able to replace the GHA functions fast with a better solution.  However, the results are now as expected.