Corrections
Now the algorithm for the sun's declination, bearing and elevation is correct.  This based upon PSA algorithm. I will implement PSA+ the next days but, however, I will do this only for a commercial version of this software. This is just the prototype. The code is quick and dirty, lot of redundancys. Not real object oriented. Looks like shit. However. Results are good so far.