Added DMS format check to all DMS fields
So far it seems that the DMS check will run without any trouble. Now users are forced to enter valid DMS format in all DMS fields. The major point of interest is now to calculate the proper GHA for the sun. After this next step will be to add calculation of GHA and DECLINATION for the Moon and Planets. Some work to do.