Removed the South CheckBox
Since the DMS formater is working and data from the SUN calculator is pushed to the Simple dialog the South checkbox makes no sense anymore. I added this checkbox to the application to keep it simple. It was a trick to avoid development of a working DMS formater. I do not like the way this works today in "so called" OOP programming language. All these GUI developers do not really understand OOA, OOD and OOP. Only GUI which worked perfect for OOP was Gupta/Centura SQLWindows. That's why I tried to avoid this effort. However, we now have it but it's much more a workaround than an OOP solution.