Well, Shaka, when the walls fell.
Holy Moly. I have it. Now the calculation of GHA, SHA, DECLINATION for a given Date and Time corresponds to the Nautical Almanac. Shaka! Added calculation mode for DIP and Refraction. You can take the figures from Nautical Almanac or those calculated by this software. Both works. Hot fix for a fatal error came with some changes to the hard wired CB data. Added tooltips to all of the inputfields and checkboxes.