New Sextant Correction Dialog
This dialog allow to compute DIP by height of navigator above sealevel while taking a shot plus the calculation of atmospheric corrections (Refraction) by the algorithms Meeus published in his book "Astronomical Algorithm's". It's also possible to enter here values from the Nautical Almanac or the "Longterm Almanac". It seems that the calculated values are also good. I compared them with the Nautical Almanac and in my opinion they are much better.