Well, Shaka, when the walls fell.
Holy Moly. I have it. Now the calculation of GHA, SHA, DECLINATION for a given Date and Time corresponds to the Nautical Almanac. Shaka! Added calculation mode for DIP and Refraction. You can take the figures from Nautical Almanac or those calculated by this software. Both works. Hot fix for a fatal error came with some changes to the hard wired CB data. Added tooltips to all of the inputfields and checkboxes. Fixed TZ issue. My recommendation is: User a watch at TZ 0 and set System Time of your Android to GMT. I made one huge mistake because the emulator run at GMT and I tried to  get correct figures for local time. Fixed also a problem which occures when the sun is next to Equator. You saw this error (for example) 20.03.2023 from 21:00 to 22:00 O'clock when the sun moved above the equator.  Bug in Real2DMS function.